+++ src/components/agentDebugger/index.ts
@@ -1,0 +1,2 @@
+export { default } from './AgentDebuggerPanel';
+export { AgentDebuggerToggle } from './AgentDebuggerToggle';